Efficiency tweaks
Fixed a bug with strict tolerance and peak changes.

Improved efficiency of the clear_smoothing_values function (now it only
checks for cyan individuals)

Improved efficiency for changing smoothing parameter - now when sp
limits are turned on, only the functions will change that have smoothing
parameters outside of the range.